+++ Changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [unreleased]
+- Allow inactive groups in the graders table to be toggled for display (#6778)
+- Enable plotly rendering for jupyter notebooks (#6871)
 - Added new API routes for the `create` and `destroy` actions of SectionsController (#6873)
 
 ## [v2.4.1]